nixos/nvidia: optionally disable setup commands for reverse sync
This adds an option to disable the display manager setup commands for PRIME reverse sync.

In some cases, having the display manager set up the providers causes Xorg to crash when a user attempts to reconfigure the monitor layout using xrandr after logging in.

The option is enabled by default to preserve backwards compatibility.